Move render_neurons and template_spaces into package
