+++ colormips/fiji_plugin_output/segid648518346494405175_in_JRC2018_VNC_FEMALE_461_0256_DSLT01.0519_thre03.0007.tif
