+++ colormips/fiji_plugin_output/segid648518346494405175_in_JRC2018_VNC_MALE_461_0256_DSLT01.3419_thre02.9558.tif
