+++ colormips/fiji_plugin_output/segid648518346494405175_in_JRC2018_VNC_UNISEX_461_0256_DSLT01.2373_thre03.1428.tif
