+++ colormips/segid648518346494405175_in_JRC2018_VNC_UNISEX_461.nrrd
