+++ fanc/render_neurons.py
@@ -5,39 +5,15 @@ import os
 
 import numpy as np
 import tqdm
-
 import navis
 import flybrains
-from caveclient import CAVEclient
 from meshparty import trimesh_io
-from FANC_auto_recon.transforms.template_alignment import warp_points_FANC_to_template
-import npimage # pip install git+https://github.com/jasper-tms/npimage
+import npimage
 import npimage.graphics
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))  # Make sure template_spaces' folder is on path
-from template_spaces import template_info, get_nrrd_metadata
-
-
-show_help = """\
-    Render a neuron that has been reconstructed in FANC into an image stack aligned to a VNC template.
-    The rendered neuron stack can be used for generation of a depth-colored MIP for use in mask searching.
-
-    Usage:
-      ./render_neuron_into_template_space.py seg_id [name_of_template_space]
-
-    seg_id must be a segment ID from the FANC production segmentation.
-    name_of_template_space must be the name of a template space. The options
-      for this can be found in FANC_auto_recon/lm_em_comparisons/template_spaces.py
-      If omitted, a default of 'JRC2018_VNC_UNISEX_461' will be used.
-
-    Example usage:
-      ./render_neuron_into_template_space.py 648518346494405175
-    Output file will be named segid648518346494405175_in_JRC2018_VNC_UNISEX_461.nrrd
-
-    WARNING: This script can take 10+ minutes to run on large neurons, as
-      large neurons can have meshes with many millions of faces.
-      A small neuron (~100k faces) that can be used for testing is 648518346516214999
-"""
+from . import auth
+from .transforms.template_alignment import warp_points_FANC_to_template
+from .template_spaces import template_info, get_nrrd_metadata
 
 
 def render_mesh_into_template_space(seg_id: int, target_space: str, header_pixels=0):
@@ -49,7 +25,7 @@ def render_mesh_into_template_space(seg_id: int, target_space: str, header_pixel
     target_info = template_info[target_space]
 
     # Setup
-    client = CAVEclient('fanc_production_mar2021')
+    client = auth.get_caveclient()
     mm = trimesh_io.MeshMeta(
         cv_path=client.info.segmentation_source(),
         disk_cache_path=os.path.expanduser('~/.meshes'),
@@ -110,22 +86,3 @@ def render_mesh_into_template_space(seg_id: int, target_space: str, header_pixel
                  'segid{}_in_{}.nrrd'.format(seg_id, target_space),
                  metadata=get_nrrd_metadata(target_space),
                  dim_order='xyz')
-
-
-def main():
-    if len(sys.argv) == 1:
-        print(show_help)
-        return
-
-    seg_id = int(sys.argv[1])
-
-    if len(sys.argv) >= 3:
-        template_space_name = sys.argv[2]
-    else:
-        template_space_name = 'JRC2018_VNC_UNISEX_461'
-
-    render_mesh_into_template_space(seg_id, template_space_name)
-
-
-if __name__ == '__main__':
-    main()
