+++ fanc/template_spaces.py
